+++ tests/testthat/test-parse.R
@@ -99,6 +99,8 @@ test_that('parse uri functions okay', {
   )
 })
 
+
+# tags ----
 # do we match:
 # https://github.com/bluesky-social/atproto/blob/main/packages/api/tests/rich-text-detection.test.ts
 test_that('parse tags inline matches bluesky typescript', {
@@ -297,5 +299,34 @@ test_that('parse tags inline matches bluesky typescript', {
   expect_equal(
     parse_tags('no match 1?: #1?'),
     list(list(list()))
+  )
+
+# emoji ----
+test_that('parse emoji functions okay', {
+  skip_if_not_installed('emoji')
+
+  expect_equal(
+    parse_emoji('Test quote with :emoji:'),
+    'Test quote with :emoji:'
+  )
+
+  expect_equal(
+    parse_emoji('Test quote with :emoji: and :fire:'),
+    'Test quote with :emoji: and 🔥'
+  )
+
+  expect_equal(
+    parse_emoji(':fire:'),
+    '🔥'
+  )
+
+  expect_equal(
+    parse_emoji('Test quote with :emoji: and :fire: and :confetti_ball:'),
+    'Test quote with :emoji: and 🔥 and 🎊'
+  )
+
+  expect_equal(
+    parse_emoji('Test quote with :emoji: and :fire: and :confetti_ball: from r package `bskyr` via @bskyr.bsky.social (https://christophertkenny.com/bskyr/)'),
+    'Test quote with :emoji: and 🔥 and 🎊 from r package `bskyr` via @bskyr.bsky.social (https://christophertkenny.com/bskyr/)'
   )
 })