+++ NEWS.md
@@ -1,13 +1,28 @@
 # bskyr 0.2.0
 
+* Improves processing of posts into tidy objects, impacting:
+  * `bs_get_posts()`: Posts are now returned as a tibble with one row per post, regardless of type.
+  * `bs_get_author_feed()`: Posts no longer create extra columns when there are multiple embeds.
 * Adds support for starter packs (#7)
   * `bs_get_actor_starter_packs()` retrieves a list of starter packs for a specific actor.
   * `bs_get_starter_pack()` retrieves a specific starter pack.
   * `bs_get_starter_packs()` retrieves a list of starter packs.
 * Adds support for additional search parameters in `bs_search_posts()` (#6)
-* Improves list functionality
+* Adds support for emoji in the text of posts, powered by the emoji package. (#11)
+* Adds `bs_url_to_uri()` to convert a URL to a Bluesky URI.
+  * This additionally allows `bs_get_posts()` to take URLs.
+* Add support for posting videos within `bs_post()`, including gifs (#5).
+* Improves list reading functionality
   * `bs_get_actor_lists()` retrieves all lists made by an actor
   * `bs_get_list()` retrieves a view of a list
+* Expands support for working with lists (#9)
+  * `bs_new_list()` creates a new list
+  * `bs_delete_list()` deletes a list
+  * `bs_new_list_item()` adds someone to a list 
+  * `bs_delete_list_item()` removes someone from a list
+* Adds new helper function `bs_extract_record_key()` to extract the record id or key from a URL or URI.
+* Adds support for getting relationships between users with `bs_get_relationships()`.
+* Adds support for getting quote posts for a given post with `bs_get_quotes()`.
 
 # bskyr 0.1.3